+++ code/go/0chain.net/sharder/blockstore/mocks_test.go
@@ -1,7 +1,9 @@
 package blockstore
 
 import (
+	"encoding/hex"
 	"fmt"
+	"golang.org/x/crypto/sha3"
 	"log"
 	"math"
 	"os"
@@ -231,4 +233,25 @@ func mockDTierNilVolumesConfig(t testingT) *viper.Viper {
 	)
 
 	return cfg
+}
+
+func mockBWR() *BlockWhereRecord {
+	bin, _ := time.Now().MarshalBinary()
+	hash := sha3.Sum256(bin)
+	return &BlockWhereRecord{
+		Hash:      hex.EncodeToString(hash[:]),
+		BlockPath: hex.EncodeToString(hash[:]),
+		ColdPath:  hex.EncodeToString(hash[:]),
+		Tiering:   0,
+	}
+}
+
+func mockUBR() *UnmovedBlockRecord {
+	now := time.Now()
+	bin, _ := time.Now().MarshalBinary()
+	hash := sha3.Sum256(bin)
+	return &UnmovedBlockRecord{
+		Hash:      hex.EncodeToString(hash[:]),
+		CreatedAt: now,
+	}
 }
